Add methods and object for QC to PointsService.wsdl
  + Bump versions to 1.1.0-SNAPSHOT (still backwards compatible)
  + Add skeleton methods to PortImpls
  + VS Code auto-format

2 new methods exist - getBalance and setBalance. They both have async
versions( {get,set}BalanceAsync ). The other methods do not have an
async correspondent (used some XPath dark magic to achieve this).

They have as return value or argument a taggedBalance, that has 2 fields:
points and tag. taggedBalance.tag is actually just a sequence number,
because we don't think we need a client ID in this specific case to
implement a sequentially-consistent Quorum Consensus.

Changes to be committed:
	modified:   pts-contract/PointsService.wsdl
	new file:   pts-ws{-cli}/src/jaxws/binding.xml (adds Async variants)
	modified:   pom.xml
	modified:   pts-ws{-cli}/pom.xml
	modified:   pts-ws/src/.../PointsPortImpl.java
	modified:   pts-ws-cli/src/.../PointsClient.java